requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
bitarray==3.0.0
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.1
ckzg==2.0.1
cytoolz==1.0.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.1.0
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
multidict==6.1.0
parsimonious==0.10.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
python-decouple==3.8
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.6.1
websockets==13.1
yarl==1.18.3
